+++ tests/test_matcher.py
@@ -368,7 +368,7 @@ def pm_pattern_file_fmt_directory_traversal_test(pytester: pytest.Pytester) -> N
             [pytest]
             addopts = -vv -ra
             pm-patterns-base-dir = .
-            pm-pattern-file-fmt = ../{{class}}/../{{fn}}
+            pm-pattern-file-fmt = ../{class}/../{fn}
         """
       )
     # Run all tests with pytest
@@ -376,6 +376,31 @@ def pm_pattern_file_fmt_directory_traversal_test(pytester: pytest.Pytester) -> N
     result.stderr.re_match_lines([
         'ERROR: Directory traversal is not allowed for `pm-pattern-file-fmt` or `pm-patterns-base-dir` option'
       ])
+
+
+@pytest.mark.parametrize(
+    ('fmt', 'error_string')
+  , [
+        ('{unknown}', "'pm-pattern-file-fmt' has invalid placeholder: 'unknown'")
+      , ('{fn}/{unknown}/{class}', "'pm-pattern-file-fmt' has invalid placeholder: 'unknown'")
+      , ('{satu}-{dua}', "'pm-pattern-file-fmt' has invalid placeholders: 'satu', 'dua'")
+      , ('{fn', "'pm-pattern-file-fmt' has incorrect format: expected '}' before end of string")
+      , ('fn}', "'pm-pattern-file-fmt' has incorrect format: single '}' encountered in format string")
+      , ('static string', "'pm-pattern-file-fmt' should have at least one placeholder")
+    ]
+  )
+def pm_pattern_file_bad_fmt_test(pytester: pytest.Pytester, fmt, error_string) -> None:
+    # Write a sample bad config file
+    pytester.makefile(
+        '.ini'
+      , pytest=f"""
+            [pytest]
+            pm-pattern-file-fmt = {fmt}
+        """
+      )
+    # Run all tests with pytest
+    result = pytester.runpytest()
+    result.stderr.re_match_lines([f'ERROR: {error_string}'])
 
 
 @pytest.mark.parametrize(